Fixed syntax in deployment scripts
Fixed sahara-ci-syntax-check job for deployment scripts

+++ config/devstack/micro-osd.sh
@@ -1,3 +1,4 @@
+#!/bin/bash -xe
 #
 #    Copyright (C) 2013,2014 Loic Dachary <loic@dachary.org>
 #
@@ -45,7 +46,8 @@ sudo apt-get --yes install ceph ceph-common
 pkill ceph-mon || true
 pkill ceph-osd || true
 if $old; then
-   rm -rf $DIR/*
+   DIR=""
+   rm -rf "${DIR:?}/"*
 else
    rm -rf $DIR
 fi
